Support getTokenHTML for older contracts
