+++ contracts/dependency-registry/Scripty.sol
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Unlicense
+pragma solidity 0.8.17;
+
+import "scripty.sol/contracts/scripty/ScriptyBuilder.sol";
