+++ contracts/interfaces/0.8.x/IGenArt721CoreTokenHashProviderV1.sol
@@ -1,0 +1,5 @@
+pragma solidity 0.8.17;
+
+interface IGenArt721CoreTokenHashProviderV1 {
+    function tokenIdToHash(uint256 tokenId) external view returns (bytes32);
+}
